move snapshot tests
