+++ src/magic_folder/test/test_magic_folder.py
@@ -2471,109 +2471,3 @@ class RealTestAliceBob(MagicFolderAliceBobTestMixin, AsyncTestCase):
         d = super(RealTestAliceBob, self).setUp()
         self.inotify = get_inotify_module()
         return d
-
-class TahoeSnapshotTest(AsyncTestCase):
-    """
-    Tests for the snapshots
-    """
-    @defer.inlineCallbacks
-    def setUp(self):
-        """
-        Create a Tahoe-LAFS node which contain some magic-folder configuration
-        and run it.
-        """
-        yield super(TahoeSnapshotTest, self).setUp()
-        self.client_fixture = SelfConnectedClient(reactor)
-        yield self.client_fixture.use_on(self)
-
-        self.tempdir = self.client_fixture.tempdir
-        self.node_directory = self.client_fixture.node_directory
-
-
-    @defer.inlineCallbacks
-    def test_create_new_tahoe_snapshot(self):
-        """
-        create a new snapshot (this will have no parent snapshots).
-        """
-
-        # Get a magic folder.
-        folder_path = self.tempdir.child(u"magic-folder")
-        folder_name = b"default"
-        folder_alias = b"magik:"
-        outcome = yield cli(
-            self.node_directory, [
-                b"create",
-                b"--name", folder_name,
-                folder_alias,
-                b"test_snapshot",
-                folder_path.asBytesMode().path,
-            ],
-        )
-
-        self.assertThat(
-            outcome.succeeded(),
-            Equals(True),
-        )
-
-        os.mkdir(folder_path.asBytesMode().path)
-        file_path = folder_path.child("foo")
-        file_path.touch()
-
-        from twisted.internet import reactor
-        treq = HTTPClient(Agent(reactor))
-
-        snapshot = TahoeSnapshot(node_directory=self.node_directory, filepath=file_path)
-
-        snapshot_uri = yield snapshot.create_snapshot([], treq)
-
-        self.assertThat(snapshot_uri, StartsWith("URI:DIR2-CHK:"))
-
-    @defer.inlineCallbacks
-    def test_snapshot_extend(self):
-        """
-        create a new snapshot and extend it with a new snapshot.
-        """
-
-        # Get a magic folder.
-        folder_path = self.tempdir.child(u"magic-folder")
-        folder_name = b"default"
-        folder_alias = b"magik:"
-        outcome = yield cli(
-            self.node_directory, [
-                b"create",
-                b"--name", folder_name,
-                folder_alias,
-                b"test_snapshot",
-                folder_path.asBytesMode().path,
-            ],
-        )
-
-        self.assertThat(
-            outcome.succeeded(),
-            Equals(True),
-        )
-
-        os.mkdir(folder_path.asBytesMode().path)
-        file_path = folder_path.child("foo")
-        file_path.touch()
-
-        from twisted.internet import reactor
-        treq = HTTPClient(Agent(reactor))
-
-        snapshot = TahoeSnapshot(node_directory=self.node_directory, filepath=file_path)
-
-        initial_snapshot_uri = yield snapshot.create_snapshot([], treq)
-
-        self.assertThat(initial_snapshot_uri, StartsWith("URI:DIR2-CHK:"))
-
-        # write something to the file
-        with open(file_path.asBytesMode().path, "w") as f:
-            f.write("foobar")
-
-        snapshot_uri = yield snapshot.create_snapshot([initial_snapshot_uri], treq)
-
-        self.assertThat(snapshot_uri, StartsWith("URI:DIR2-CHK:"))
-
-        # the returned URI should be different from the first snapshot URI
-        self.assertNotEqual(snapshot_uri, initial_snapshot_uri)
-
